every CLI command has its own module, CLI commands are now also extensions
