+++ test/cli/helpers.py
@@ -6,7 +6,7 @@ import click
 import click.testing
 
 from xcube.api.new import new_cube
-from xcube.cli.cli import cli
+from xcube.cli.main import cli
 from xcube.util.dsio import rimraf
 
 TEST_NC_FILE = "test.nc"
